requirements.txt
Django
PyJWT
simplejson